Refactored CustomElements import/export step.
